refactor: update user service to new proto
