+++ user-service/src/controller/user_crud_service_handlers/get_user_handler.ts
@@ -1,6 +1,6 @@
-import { GetUserRequest, GetUserResponse } from '../../proto/user-service';
+import { GetUserRequest, GetUserResponse } from '../../proto/user-crud-service';
 import { IApiHandler, ApiRequest, ApiResponse } from '../../api_server/api_server_types';
-import { IStorage, IUserStore } from '../../storage/storage.d';
+import { IStorage, IUserStore } from '../../storage/storage';
 import { convertStoredUserToPasswordUser } from '../../model/user_helper';
 
 function getHeaderlessResponse(resp: GetUserResponse): ApiResponse<GetUserResponse> {
