+++ user-service/src/controller/user_service_handlers/logout_handler.ts
@@ -1,4 +1,4 @@
-import { LogoutErrorCode, LogoutRequest, LogoutResponse } from '../../proto/user-bff-service';
+import { LogoutErrorCode, LogoutRequest, LogoutResponse } from '../../proto/user-service';
 import { IApiHandler, ApiRequest, ApiResponse } from '../../api_server/api_server_types';
 import { IAuthenticationAgent } from '../../auth/authentication_agent_types';
 import Constants from '../../utils/constants';
