+++ user-service/src/index.ts
@@ -3,8 +3,8 @@ import { Request, Response } from 'express';
 import getApiServer from './api_server/api_server';
 import createAuthenticationService from './auth/authentication_agent';
 import { IAuthenticationAgent } from './auth/authentication_agent_types';
-import UserBFFServiceApi from './controller/user_service_bff_controller';
-import UserServiceApi from './controller/user_service_controller';
+import UserBFFServiceApi from './controller/user_service_controller';
+import UserServiceApi from './controller/user_crud_service_controller';
 import AppStorage from './storage/app_storage';
 import loadEnvironment from './utils/env_loader';
 import Constants from './constants';
